feat: #1375 — @harmony/shared package を新設し DraftResourceType の frontend/backend 重複定義を解消
PR #1372 → #1376 (#1374) で backend 内 4 表現 (型 union / WS Set / MCP allowlist / MCP tool schema enum) の
drift を `DRAFT_RESOURCE_TYPES` (as const) 1 配列起点に集約した次段階。frontend 側
`frontend/src/types/draft.ts` に残っていた手書き union 型を `@harmony/shared` package 経由で
backend と同じ source-of-truth に統合し、新 resource type 追加が **配列 1 箇所更新で完結** する状態を
確立する。

## 変更内容

### 1. `@harmony/shared` package 新設 (`shared/`)
- `shared/package.json`: npm workspaces 内 package `@harmony/shared`、private、ESM
- `shared/tsconfig.json`: ES2022 + bundler resolution、`dist/` に declaration + js を emit
- `shared/src/draftResourceTypes.ts`: `DRAFT_RESOURCE_TYPES` (as const) + `DraftResourceType`
  (typeof[number]) — 14 種類の resource type を一括定義
- `shared/src/index.ts`: re-export entry
- `prepare` script で npm install 時に自動 build (`shared/dist/index.js` + `.d.ts`)
- `shared/.gitignore`: `dist/` を ignore (build artifact)

### 2. top-level `package.json` を npm workspaces 化
- `"workspaces": ["shared", "frontend", "backend"]` 宣言
- `install:all` を `npm install` 一発に簡略化
- per-workspace lock (`frontend/package-lock.json`, `backend/package-lock.json`) を削除し、
  top-level `package-lock.json` 1 本に統合
- 副次効果として hoisting 整理により workspaces 全体で重複 install が解消

### 3. backend `editSessionStore.ts` を `@harmony/shared` import に置換
- 旧: `DRAFT_RESOURCE_TYPES` を本ファイルで定義 + 型導出
- 新: `@harmony/shared` から import + 互換性維持のため re-export
- `VALID_RESOURCE_TYPES = new Set(DRAFT_RESOURCE_TYPES)` は backend の runtime allowlist として
  本ファイルに残し、WS / MCP handler の import 経路は維持

### 4. frontend `types/draft.ts` を re-export-only に変更
- 旧: 手書き string literal union 14 種類
- 新: `export { type DraftResourceType } from "@harmony/shared"` のみ
- frontend 30 箇所の `import type { DraftResourceType } from "../types/draft"` は無変更で動作

### 5. eslint-plugin-react-hooks を `7.0.1` exact pin
- npm workspaces lock 再生成で `^7.0.1` → 7.1.1 に semver 解決されると別 file の
  pre-existing React Compiler 警告 3 件が error 化するため、現状動作維持目的で固定。
  upgrade + 解消は follow-up ISSUE で trace 確立予定 (鉄則 0)

## 受入条件 (ISSUE #1375)

- [x] frontend `DraftResourceType` を backend 側 source of truth と構造的に同期する経路が確立
  → `@harmony/shared` package 経由で同一 module を import
- [x] 新 resource type 追加が 1 箇所更新で frontend / backend 両 type に反映されること
  → `shared/src/draftResourceTypes.ts` の `DRAFT_RESOURCE_TYPES` 配列 1 箇所のみ更新
- [x] backend `DRAFT_RESOURCE_TYPES` 配列 1 箇所の更新で MCP tool schema / WS allowlist /
      backend type / frontend type すべてが追従する状態
  → backend の `tools.ts` (MCP enum spread) / `editSessionStore.ts` (VALID_RESOURCE_TYPES Set) /
    frontend `types/draft.ts` (re-export) すべてが `@harmony/shared` の同配列から導出

## 検証

- backend build (`tsc`): pass
- backend test (vitest): 800/800 pass
- backend dist 動作確認: `node -e "import('./dist/editSessionStore.js')"` で 14 種類取得・Set size 14 確認
- frontend build (Vite): pass
- frontend test (vitest): 3008/3009 pass (1 skipped、pre-existing)
- frontend lint (eslint): pass
- shared dist 自動 build: `prepare` hook で `npm install` 時に `shared/dist/` 生成確認

## 関連

- 親 PR: #1376 (backend 内 4 表現集約)
- 親 ISSUE: #1374
- 親メタ: #1298 / RFC #1284
- 起源 PR: #1372 (#1368 GenericDefinition save 統合)

Closes #1375
