+++ frontend/package.json
@@ -20,6 +20,7 @@
   },
   "dependencies": {
     "@dnd-kit/core": "^6.3.1",
+    "@harmony/shared": "*",
     "@dnd-kit/sortable": "^10.0.0",
     "@dnd-kit/utilities": "^3.2.2",
     "@grapesjs/react": "^2.0.0",
@@ -54,7 +55,7 @@
     "ajv-formats": "^3.0.1",
     "autoprefixer": "^10.5.0",
     "eslint": "^9.39.4",
-    "eslint-plugin-react-hooks": "^7.0.1",
+    "eslint-plugin-react-hooks": "7.0.1",
     "eslint-plugin-react-refresh": "^0.5.2",
     "globals": "^17.4.0",
     "jsdom": "^29.0.2",
