+++ shared/src/index.ts
@@ -1,0 +1,1 @@
+export { DRAFT_RESOURCE_TYPES, type DraftResourceType } from "./draftResourceTypes.js";
